Tests/LibWeb/Text/expected/indexeddb-close-during-upgrade-event-ordering.txt

upgradeneeded
upgrade tx complete
error:AbortError
